added react query for custom hooks and state context management
