+++ package.json
@@ -20,6 +20,7 @@
     "@miblanchard/react-native-slider": "^2.6.0",
     "@react-native-async-storage/async-storage": "1.23.1",
     "@react-navigation/native": "^6.0.2",
+    "@tanstack/react-query": "^5.59.20",
     "expo": "~51.0.28",
     "expo-constants": "~16.0.2",
     "expo-drizzle-studio-plugin": "^0.0.2",
